rename coloring -> routing
